third_party/WebKit/LayoutTests/external/wpt/console/console-timeline-timelineEnd-historical.any-expected.txt

This is a testharness.js-based test.
FAIL 'timeline' function should not exist on the console object assert_equals: console.timeline should be undefined expected (undefined) undefined but got (function) function "function timeline() { [native code] }"
FAIL 'timelineEnd' function should not exist on the console object assert_equals: console.timelineEnd should be undefined expected (undefined) undefined but got (function) function "function timelineEnd() { [native code] }"
Harness: the test ran to completion.
